src/main/java/com/dconnell/server/service/DefaultUserService.java
package com.dconnell.server.service;

import com.dconnell.server.model.user.User;
import com.dconnell.server.respository.UserRepository;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;

public class DefaultUserService implements UserDetailsService {

    private final UserRepository userRepository;

    public DefaultUserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public User loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userRepository.findUserByUsername(username);
        if (user == null) {
            throw new UsernameNotFoundException("Could not find user.");
        }
        return user;
    }

}